requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.6
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==2.0.0
ipdb==0.13.13
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.22
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pyotp==2.9.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.8
